Feat: 전체 조회 추가  #S11P31C108-130

See merge request s11-final/S11P31C108!64

+++ backend/kickboard/src/main/java/kb/user/internal/domain/Location.java
@@ -19,13 +19,13 @@ public class Location {
     @Id
     @Column(name = "locationId")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int locationId;
+    private Long locationId;
 
     @Column(name = "locationName")
     private String name;  // 구역 정보
 
     @Builder
-    public Location(int locationId, String name) {
+    public Location(Long locationId, String name) {
         this.locationId = locationId;
         this.name = name;
     }
